Created the full list of known "Ped Reset Flags" on the internet. Read commit description for more.
I've hard codded the value from 0 to 459 and only tried a few of them.
If R* at some points added in the middle since the leaked source code, some stuff might not be accurate,
but as far of my testings, everything here is correct.
There might be missing new flags since the source code leak.